updates MQ2 to be compatible with QuPath 0.5.1
